requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-environ==0.4.5
djangorestframework==3.12.4
environ==1.0
idna==2.10
psycopg2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
ultimate-sitemap-parser==0.5
urllib3==1.26.4
